requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
jinja2==3.1.4
